api/requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
marshmallow==3.20.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
pusher==3.3.2
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
Flask-Talisman==1.1.0
supabase==2.0.3
python-jose==3.3.0
bcrypt==4.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
validators==0.22.0
pytest==7.4.3
pytest-flask==1.3.0
pydantic==2.5.2
uuid==1.30
